GIGI/requirements.txt
streamlit
sentence-transformers
torch
requests
python-dotenv
